Added a blacklist to the File Manager.  Currently the blacklist, in defaults.conf, prevents course.conf and simple.conf from being edited unless the user has administrator priviledges.
-  Also restricted users so they cannot give themselves permissions higher than they currently have.
-  Also changed "add administrators to new course" in Course Admin to add administrators as admin.

The use case is that you are added as an admin user when you create the course from Course Admin and that will give you the ability to set professors to admin status if you feel they should edit the course.conf files.  Alternatively you can just override the blacklist directly.