Clean up error messages.
These error messages are aimed at making it easier to set up the initial communication between WeBWorK and the
LMS calling WeBWorK via LTI.

Also commented out some preliminary code to be used at Rochester for deducing sections from the Blackboard course_label at
UR.

A lot of the authentication procedure is as described in the comments and specific to the needs of IU.  A future update
should breakout the process of assigning recitations and sections into a callback subroutine so that it is easier to
program in different behaviors for handling this task.  This will do for now.