Add timeout to the check_url application (lwp-request in this case).
alias(url) is run through check_url and if this
results in a hung call the timeout will force check_url to report a broken link before
the problem times out.  (Otherwise the error message will be
"Problem timed out, check for infinite loops" which is misleading. )

The current defaults are url search timeout is 40seconds, problem timeout is 60 seconds.
They are hard coded  in separate locations.