Fixed bug in instructorXMLHandler
The {text} output from the renderProblem command and from other commands must be handled differently. For the renderProbelm command this is printed directly (to the screen).  For all the other commands the {text} output (which is not usually empty) is included in a JSON hash which is then returned as output.  I added an explicit flag that alerts the content() subroutine as to whether it is handling a renderProblem command or one of the other commands.