Add helpful error message if someone uses the key word password instead of course_password
There is a site_password -- which keeps intruders from even attempting to log in to the site -- although this isn't being checked yet.
I thought course_password was better than password since it makes it clearer that it is the password associated with the courseID
and the userID.  The userID must be in courseID and have a matching course_password in order to use the site.