Fix artifact that {"server_response:"   artifact is occurring when using rerandomize from the library
This is a hack because instructorXMLRPC and renderViaXMLRPC and some of the javascripts make different assumptions
about where the output of the call is to be found. It's also not uniform between renderProblem and other calls.
This is another case where I need to write down an API and then retrofit all the different
scripts to obey it.