Moved more configuration to the ww_credentials.dist file. Check that configurations can be made in that file and that the lack of a configuration is either reported or replaced with a sensible default. (The webwork displayMode defaults to "MathJax" for example.)
The hash output modes print directly to STDOUT -- one can pipe this through less if desired, or use the output in otherwise. The output continues to be pretty printed.

It is possible to set the log file path using --log=path_to_log_file.  The path defaults to ...../webwork2/DATA/xmlrpc_results.log