handle corrupt previous answers in decodeAnswers
We got a report that a user was unable to open a problem they had previously
attempted and they were receiving this error instead:

	Can't use an undefined value as an ARRAY reference at
	/opt/webwork/webwork2/lib/WeBWorK/Utils.pm line 871.

The error refers to decodeAnswers, which is trying to use the result of a
Storable::thaw call without verifying that it succeeded (it returns undef on
some errors).

This change makes decodeAnswers catch those errors.