Changed the form of a maketext construction containing two dropdown lists.
I would suggest to use this kind of construction instead of something like $r->maketext("foo").<dropdown stuff>.$r->maketext("bar") whenever possible.
The reason is that languages have different word order and sometimes it is difficult to translate parts of an English sentence into another language while preserving the order.
Using the construction I adopted here, the translator is free to change the position of the dropdown lists in the sentence.