Set specific charset on new OPL tables
If our database defaults to the utf8 character set and we run OPL-update, it
will fail creating tables.  It fails because the primary key of the new tables
is too large.  This is because the primary key includes several VARCHAR
columns, which are 3x in size if they are utf8.

This commit changes the table creating queries to force the latin1 character
set instead of using the database default.  This allows the tables to be
created and the OPL indexes loaded into the database.