add dir="auto" attribute to <td>
This adds the attribute dir="auto" to the <td> element that displays the
students original answer. This is useful in a global dir="rtl" setting.
For mathematical formulas consisting of ASCII characters this should
imply an LTR setting, while it should imply an RTL setting when the text
entered is written in a language that is written from right to left.