This is an early version of the Perl code for the function output_problem_lang_and_dir() which sets the lang and dir attributes for the HTML DIV containing the problem text to depend on the language and text direction needed for the specific problem. It accesses PG flags which carry the language text direction requested by the problem.
This early version does not yet depend on course settings which
would allow overriding the new behavior by forcing particular
settings (or forcing no settings to be made). Such overrides will
allow a course to prevent the new code from changing the prior
behavior, or allow forcing fixed values which do not depend on the
PG flags or the defaults used by output_problem_lang_and_dir()
when the PG flags are not set.

The math4 and math4-ar template files were modified in a prior
commit to call this function.

Code was added to pg/macros/PG.pl to enable setting the PG flags
on which this code depends.