Moved the code to select problem language and direction into a Util function called get_problem_lang_and_dir() which can be used by whichever ContentGenerator modules need it.
This replaces the cod which was in lib/WeBWorK/ContentGenerator/Problem.pm
with a function call which is used to generate either the HTML code, or the
options to pass into CGI::div.

Some changes were made with the functionality/defaults, in particular,
except for mode "none" it will set the value for lang/dir even if they
match the course-wide values expected.