Add a mariadb.cnf file to the Docker image to force the database to use utf8mb4. This needs to be done BEFORE the Docker database storage volume is created.
A special version called
	docker-compose.yml.specialUTF8MB4-db-storage
can be used to replace docker-compose.yml, and will
change the name of the database container, the database
named storage volume, and make related settings. Replace
docker-compose.yml with the file will force Docker to
run the containers in a manner which will not intefere with
the non-utf8mb4 database used by the prior configuration.
That approach would allow one to switch between the use of
Docker containers using a utf8mb4 database and using an
old style database, just by using the "correct" settings
in docker-compose.yml for each "case".