Changes to the XMLRPC system to allow it to handle UTF-8 encoded problems.
Rendering works already. However, submitting UTF-8 encoded
text which is not English (Latin1) as form values (in particular
as answer) is not working yet. Due to this, the values set on the
submit buttons are currently forced back into English.

clients/sendXMLRPC.pl was extended to take a language setting
as an argument.

lib/WebworkClient.pm contains code to set the HTML language and
text direction both at the level of the "page" and at the level
of the DIV element enclosing a problem (based on data set in the
PG file).

lib/WeBWorK/Localize.pm received a new method getLangHandle()
used by lib/WebworkClient.pm to get access to maketext.

lib/WeBWorK/Utils/DetermineProblemLangAndDirection.pm was extended
to allow forcing options not available via a request hash when
the XMLRPC system is rendering problems.
Small related changes and some cleanup work was done to several of
the formats in lib/WebworkClient/ including changes to support UTF-8,
to standardize the title setting, to the copyright dates. Additional
settings were added to the <form> element and to the "problem-content"
DIV element to be more similar to what is generated by the regular
ContentGenerator system. "forcePortNumber" was added as a new hidden
field. The value settings of the submit buttons are set by variables
so they can get localized text. At present, this is not in active use
as UTF-8 values render properly but then cause problems when the
buttons are clicked and the non-English values are sent as form data.

This PR also includes changes to allow forcing a port number to be
used by the XMLRPC system. This is intended to help when using the
XMLRPC system with a Docker container or a host which requires using
a port number for the site_url and form_action_url. This is managed
by setting the value of forcePortNumber in the credentials and/or
in the address used to access html2xml. Ex:

http://localhost:8080/webwork2/html2xml?&answersSubmitted=0&language=en&sourceFilePath=some_pg_file.pg&problemSeed=123567890&displayMode=MathJax&courseID=daemon_course&userID=daemon&course_password=daemon&outputformat=simple&forcePortNumber=8080