Some changes to how a set definition file (SDF) is imported (by lib/WeBWorK/ContentGenerator/Instructor/ProblemSetList2.pm).
1. If the file contains a reducedScoringDate whose value contains either
"12/31/1969" or "01/01/1970" (as a substring) then it will always be treated
as being Unix epoch 0. This is intended to solve problems when during the
import a time-zone issue interferes with reading in dates which correspond
to Unix epoch 0. However, it will also catch any other time on those dates
and also treat is as Unix epoch 0. Since no reasonable SDF would have other
times on those dates as the value of reducedScoringDate, the somewhat overly
inclusive behavior should not cause problems.

2. Until this commit, when reducedScoringDate parsed to Unix epoch 0, the code
would replace the value by an automatically calculated value based on the
course setting for reducedScoringPeriod regardless of whether the SDF had
enableReducedScoring set to "Y". This commit prevents to override of Unix epoch
0 by that code unless the file sets enableReducedScoring to "Y". That behavior
will leave the reducedScoringDate as Unix epoch 0 as would typically be the
case in a course where there is no use of reduced scoring.