Add/modify comments in 	lib/WeBWorK/ContentGenerator/Feedback.pm to emphasize that we are making the $sender RFC 2047 compliant to get support for Unicode full names. The modified version of rfc822_mailbox() applies automatically to the feedback recipients, and a comment was added about that. Finally, add a comment that in the future the $subject may need similar handling, if non-ASCII Unicode characters can be included in any of the fields used to generate the subject line.
Add / edit explanatory comments in
	lib/WeBWorK/DB/Record/User.pm
and also refer to the future issue of supporting Unicode in "international
emails" addresses and headers: RFC 6531 and RFC 6532. The support for those
features are not yet widely adopted in the existing SMTP mail infrastructure.

The subroutine rfc822_mailbox() has been kept with the existing name, even
though it is now RFC 2047 compliant.