Changes to support multilingual PDF generation.
The main goal was to support multilingual PDF generation using
XeLaTeX and polyglossia. My test case was English+Hebrew, so I
have added XeLaTeX conf/snippets/hardcopyThemes directories for
using polyglossia for English and English+Hebrew. Fonts (culmus)
were installed in the Docker image to provide the Hebrew fonts
needed to test the English+Hebrew xelatex compilation.

conf/localOverrides.conf.dist contains sample code to use the\
new themes.

conf/site.conf.dist contains a commented out option to use
xelatex instead of pdflatex.

The docker-config/ has additional sample config files, and
docker-entrypoint.sh was moved into docker-config/.
In particular there are very rough sample files for using
WW with SSL in Docker, but the placeholder certificate and
key files would need to be put in place, as well as various
settings adjusted in the config files.

The code in lib/WeBWorK/ContentGenerator/Hardcopy.pm which displays
the available hardcopyThemes options was changed to avoid the
use of maketext. Local installations may have local hardcopyThemes
and they will not have entries in the PO translations files, and
the prior code would trigger errors as a result.

The version of xelatex and various latex packages in Ubuntu 16.04
did not manage to compile the Hebrew+English latex files I tested,
but they compile under Ubuntu 18.04. As a result, the Dockerfile was
changed to build upon Ubuntu 18.04 (LTS), which had originally
been intended to be done in a different PR. Along the way, many
Perl modules are now installed from the Ubuntu repositories, which
saved on "cpanm" local installation. Several additional tools, as
well as some packages for SSL support were added to the image.

The docker-compose.yml was updated, and includes sample code to use
various external resources, and to store Apache logs outside the image.
At present, the docker config files are set up under the assumption that
the OPL is mounted from OUTSIDE the docker image. That saves on build time,
and allows sharing it between different images (but may slow things down
somewhat) but there is a commented out block in Dockerfile to put the
OPL in the image.

The Dockerfile also updates the base Ubuntu image, created locales,
sets the timezone, etc. Several lines for SSL support are commented
out, and should be enabled for sites using SSL, once the relevant
customized configuration files and certificates are provided to the
image (possibly via external mounts).

Settings for SQL password in docker-compose.yml should be effective
if they are set/exist when the MariaDB container boots up with an empty
database directory/volume. Thus, those passwords should be set BEFORE
the first "docker-compose up" or after the mysql data volume is deleted.