Add problemUUID to parameters that can be submitted from a problem form.
problemIdentifierPrefix is a (deprecated) synonym which can also be used in a form -- it is renamed to problemUUID in the environment.

The $envir{problemUUID} value is determined from the input form problemUUID if it exists, ELSE
from the input form problemIdentifierPrefix if it exists ELSE
$envir{problemUUID} might have value determined from some other source than the input form ELSE
the value of $envir{problemUUID} is set to 0.