Replace $XML_URL with $SITE_URL which is a more accurate description
The contents of $SITE_URL should be 'https://mysite.edu'  which was also the intended contents
of $XML_URL.    Because of a misleading description the instructorXML file contained 'https://mysite.edu/mod_xmlprc' which led
to xmlrpcCalls of  'https://mysite.edu/mod_xmlrpc/mod_xmlrpc'.  Fortunately the last segment was ignored so this bug remained
undiscovered for a decade or more.  It came to light while researching another bug in the configuration of the xmlrpcCall path.