Remove the setting of the hasDatepicker class from the onblur hanldler. It is certain that adding the hasDatepicker class is not what should be happening there.  The date fields already have the class, and other text inputs shouldn't get it.
Instead implement better handling of the 'changed' class for the
datepicker input fields.  If a date is changed from its original value
the input shows changed, and if it is changed back to its original value
it shows as not changed.  The code is removed from perl and implemented
in datepicker.js.