Code to solve problems XMLRPC had with UTF-8 in some places.
See: the html2xml part of the report at
	https://github.com/openwebwork/pg/issues/443
which seems to be resolved by this change.

The main change is to encode UTF-8 data which before it gets sent
to XMLRPC and decode it on arrival.

Problems which did not work over html2xml / XMLRPC before this
change are now working, and no adverse issues have been seen in
several weeks of use (over 8000 scores earned over html2xml using
the LTI modifications) on a server running this code to get the
html2xml system working with problems several of which did not load
over html2xml without this patch.