Fix additional typos.
There is a commented out line:
	# warn "$encoded_last_answer_string", $encoded_last_answer_string;
which was used to test the the answer string was being encoded correctly -- it was
however if the answer string is long, as with a picture included in the answer string, then it may be
too long to fit in the problem->last_answer field. The field needs to be changed from TEXT to LONGTEXT or BLOB or something like
that -- still trying to decided the best option.