Update the javascript for the gateway quiz module.
This utilizes data attributes to pass in the timing inforamtion, rather
than a defunct form with hidden inputs.  The timer is placed much more
appropriately.

Instead of using a native alert for notifications near the end of a
test, use bootstrap toasts.  The native alerts interupt a student's
work, forcing the student to close them.  Boostrap toasts sit quietly at
the top, and the student can continue to enter answers while the toast
is still visible.  The toasts are set to stay open until the next alert
message, although the student can close the toast earlier.

Add a quiz auto submission at the end of the quiz that occurs
approximately 10 seconds before the end of the grace period.