Fix the problem urls in the emails sent using the "E-Mail Instructor" feedback page.
This was introduced in pull request #986.

That pull request switched from obtaining the set and problem ids cached
in the WeBWorK::ContentGenerator::Problem object to attempting to find
the set and problem ids from the url parameters.  The problem is that
the feedback page does not have those url parameters.