Update the mathquill javascript.
The main change is just pulling in upstream improvements.
One notable fix is that you can now cut and paste into a text block.
Previously this resulted in errors in the console and unexpected
behavior in the mathquill input.

I have also included a minor tweak to the text output that is sent to
webwork.  That is when an operator name like arcsec is input, the text
will no longer include 'operatorname{}'.