Move the definition of the TEXINPUTS environment variable out of webwork.apache2.4-config.dist and just define it directly in Hardcopy.pm.
The previous method works, but this makes it so that things don't break
if a sysadmin does not update the webwork.apache2.4-config file.

On the other hand, if one had a desire to override the default location
then one would now have to edit Hardcopy.pm.

So the decision as to whether or not this is the way to go comes down to
whether or not this should be configurable or not.