In the pg code the ADD_CSS_FILE and ADD_JS_FILE usage is now consistent. Both methods accept an additional argument that determines if the first argument is a file located in the htdocs directory or an external url. The documentation is updated there.
Here the application of the the generated lists takes that additional
argument into account.  Currently that argument is not used for files
added from a course.conf (or other settings file).  It is assumed those
are in htdocs.  Also if the addtional argument is not provided or is 0
(or falsy) then it is also assumed the file is in htdocs.