Change the field lengths of user_id and set_id back to 100 characters (from 80 characters), as they were before WeBWorK-2.15, when they were reduced due to key length limits and the change to support utf8mb4.
This is intended to minimize problems with long set_id names and long
user_id values.

Note that both these fields are defined as TINYBLOB in all other tables
(lib/WeBWorK/DB/Record/*.pm files) in which they appear. TINYBLOB allows
255 bytes, which depending on the actual characters (as UTF-8 encoding
uses between 1 and 4 bytes per character) can be more than the 100 characters
the VARCHAR(100) permits. That is not a new issue.

In order to support this, while remaining under the key length limit,
the course_id field was shortened to VARCHAR(40). Since this field is
really of no value as this is a per-course table, any data-loss due to
possible truncations is really not important. Hopefully the truncation
will not cause problems when the field is being checked using a "where"
condition.