Attempt to prevent problems when the code might try to set a courseID which is too long = over maxCourseIdLength characters.
Prevent bin/addcourse from attempting to create such a course.

Admin course - issue error messages if trying to use a courseID which is
too long:
  - when adding a new course,
  - when renaming an existing course, and
  - when unarchiving a course to a new courseID.

Add a length tests to create() of lib/WebworkWebservice/CourseActions.pm.

Also do such length tests in lib/WeBWorK/Utils/CourseManagement.pm
in the addCourse, renameCourse, and unarchiveCourse subroutines.

Note: The _unarchiveCourse_move_away subroutine which attempts to rename
a course to "${courseID}_tmp" before an unarchive was not changed at
present. There does not seem to be a simple, quick, and reasonable fix if
the temporary name currently in use would be too long. A possible fix for
future implementation would be to select a random temporary name which is
not currently in use.