Past answer table: resize fields + limit allowed length of a set name accordingly (#1431)
* Change the field lengths of user_id and set_id back to 100 characters
(from 80 characters), as they were before WeBWorK-2.15, when they were
reduced due to key length limits and the change to support utf8mb4.

This is intended to minimize problems with long set_id names and long
user_id values.

Note that both these fields are defined as TINYBLOB in all other tables
(lib/WeBWorK/DB/Record/*.pm files) in which they appear. TINYBLOB allows
255 bytes, which depending on the actual characters (as UTF-8 encoding
uses between 1 and 4 bytes per character) can be more than the 100 characters
the VARCHAR(100) permits. That is not a new issue.

In order to support this, while remaining under the key length limit,
the course_id field was shortened to VARCHAR(40). Since this field is
really of no value as this is a per-course table, any data-loss due to
possible truncations is really not important. Hopefully the truncation
will not cause problems when the field is being checked using a "where"
condition.

* Standardize to use tab base indentation + other white-space cleanup

* Edit the comment about $maxCourseIdLength in conf/defaults.config.

* Limit set name to 100 characters, max supported by past_answer table

* Attempt to prevent problems when the code might try to set a courseID
which is too long = over maxCourseIdLength characters.

Prevent bin/addcourse from attempting to create such a course.

Admin course - issue error messages if trying to use a courseID which is
too long:
  - when adding a new course,
  - when renaming an existing course, and
  - when unarchiving a course to a new courseID.

Add a length tests to create() of lib/WebworkWebservice/CourseActions.pm.

Also do such length tests in lib/WeBWorK/Utils/CourseManagement.pm
in the addCourse, renameCourse, and unarchiveCourse subroutines.

Note: The _unarchiveCourse_move_away subroutine which attempts to rename
a course to "${courseID}_tmp" before an unarchive was not changed at
present. There does not seem to be a simple, quick, and reasonable fix if
the temporary name currently in use would be too long. A possible fix for
future implementation would be to select a random temporary name which is
not currently in use.