WW 2.16 - mods to admin page for collecting site info / security group signup (#1405)
* Add a section to the admin course main page about signing up
for the WW-security-announce Google group / mailing list.

Provide a Google form based site registration form + modified text.

(This was recreated as a new commit off a current copy of WeBWorK-2.16
and is intended to replace
https://github.com/taniwallach/webwork2/commit/aceb03f75452df2c50c70a0bea6f6c52bebcbf9a
)

* Rework registration section. Format as left aligned, do not try to
put the form into an iFrame, just give a link. Put all needed URLS
into variables in conf/defaults.config and use those values. Changes
to the text of the registation banner.

Change a few other URLS in conf/defaults.config to use https.

* Drop text about how the registration banner is hidden.