Switch to having PG insert mathquill responses.  A new translation option is passed to PG that notifies PG that MathQuill is enabled. Webwork2 (or other front end) just needs to tell PG that mathquill is in use via this translation option.  This is also used on the PG side to modify parsing behavior to work better with MathQuill.  This parsing is only in effect when parsing a student answer (and of course only when MathQuill is enabled).
The mathquill latex answers are now stored in KEPT_EXTRA_ANSWERS instead
of as an additional response appended to the answer, so the
ProblemUtil.pm create_ans_str_from_responses method no longer needs to
worry about skipping these for past answers.

Install mathquill via npm from my github branch of my mathquill fork
that contains the changes needed to make the new parsing work.  Note
that before this is merged, that fork should be merged into the main
branch of openwebwork/mathquill and the htdocs/package.json install
updated to use that branch instead.

Note that the package-lock.json file is upgraded to lockfile version 2.
This is due to having a newew version of npm installed (as needed for
ww3 development).  If this is a problem I will see what I can do about
reverting to lockfile version 1.