Clean up the way that the MathQuill toolbar is added and removed when a MathQuill input is focused and blurred.
Make the MathQuill toolbar responsive to the window height.  This makes
the toolbar work better for short windows, for example on a mobile
device or the case of a short problem in an iframe.

Make the button separated and give the container a border and background
to make the appearance more uniform and clean.

Change the toolbar buttons to a dark color.

Switch to using bootstrap tooltips instead of jquery-ui tooltips.