Update to bootstrap version 5.
The existing themes are updated, and math4-yellow theme is added.  The
point of adding the yellow theme is to demonstrate how to implement a
theme with a light primary color.  (The colors are the MWSU colors
though, and this is the theme I will be using!)

Theming is now done quite differently.  Instead of modifying the
override files (which still exist and are not updated), copy an existing
theme directory and modify the files _theme-colors.scss and
_theme-overrides.scss.  Then run `npm run gen-theme-css` from the htdocs
directory.  Note that the override files may still be used if desired,
but that approach will take a lot more work to get a nice theme.

You can also run `npm run watch-theme-dir` from the htdocs directory to
automatically compile the css files anytime one of the scss files
changes.  This is convenient when creating/modifying a theme.  Note that
`npm run gen-theme-css` will generate compressed css files with no
source maps, while `npm run watch-theme-dir` will not generate
compressed css and will generate source maps.