Add a pg htdocs location, and move the pg related files in htdocs there.
This works by adding an apache2 rewrite.  When a file in the
`$webwork_htdocs_url` (`/webwork2_files`) is requested, then the server
first checks for the file in `$webwork_htdocs_dir`.  If the file is
there, then that file is served.  If not it tries to serve the file from
`$pg_dir/htdocs`.  The url for that location is `/pg_files`.

This is done in conf/webwork.apache2.4-config.dist, so make sure to
update your copy of that file.

The relevant files in the webwork2 files used purely by pg problems have
been moved to the pg/htdocs directory.  They are deleted in this pull
request, and added in a corresponding pg pull request.  This includes
many of the javascript and css files in htdocs/js/apps, and the html
files in htdocs/helpFiles that are used for student help.

The ADD_JS_FILE and ADD_CSS_FILE parsing methods in Problem.pm,
GatewayQuiz.pm, and FormatRenderedProblem.pm have been updated to check
for the existence of files in the pg htdocs directory as well.
Furthermore, the ADD_CSS_FILE parsing method has been updated to
preserve the order that css files are added in the same way that
javascript order is preserved.

Note that the 3rd party jsxgraph library is now installed via npm, and
the files for that deleted from this repository.

Also the knowl.js and knowlstyle.css files are moved to
htdocs/js/apps/Knowl in the pg htdocs directory and the locations
updated in the files here.