Add a missed margin on one of the attempts table messages.
Also add a matching margin-bottom to `<p>` tags in problems.  Although
bootstrap adds this margin-bottom, depending on how a problem is
rendered this may not get set.

Add the `attemptResultsHeader` class to the attempt results table and
tweak the styles for that a bit.  This doesn't change the appearance for
webwork2.  This is really for the standalone rendere and webwork3 and to
keep the appearance the same in all cases.