Only show current achievement in chains.
An achievement chain is any achievement that has the same starting
string up to the first _ character, the same category, has a
max_counter > 0 and whose number is one greater than the previous
achievement. These are for achievements such as do 10,20,30,40,...
of something.

This only shows the current achievement that is being worked on
and hides all the future achievements until the previous ones
have been earned.