Display delta level achievement progress.
This adds the total number of achievement points earned and
the points needed for the current level to the achievements page.
In addition this adds the `prev_level_points` to the `$globalData`
hash in the database, so the current level progress is displayed
as the number of points earned since the previous level instead
of always showing progress bar starting from zero.