Fix the WebworkSOAP endpoint.  The POD is used in an invalid way by the Pod::WSDL library to make this work.  That library has bad idea written all over it.
I was at least able to fix the POD so that it does not emit errors when
POD generation occurs.

Note that if the WebworkSOAP endpoint is used that the hardcoded
soap_authen_key on line 37 of lib/WebworkSOAP.pm should be changed to
something more secure.  That should probably be something read from a
configuration file instead of hardcoded there.