Add the capability for students to add attachments to feedback emails.
This switches to using the perl Email::Stuffer module, instead of
directly using the Email::Simple module.  The Email::Stuffer is much
easier to work with for adding attachments.  It uses Email::Simple and
Email::MIME under the hood.

This also adds the option `$mail{maxAttachmentSize}` in site.conf which
is the maximum allowed attachment file size in megabytes.  The default
value for this is 10 megabytes.

The allowed file types are images with the extension jpg, jpeg, png, or
gif, pdf files, zip files, txt files, and csv files.  This is hardcoded
for now.  Should videos be allowed?