Create ProblemSet.pm page for gateway sets.
This makes it so all problem sets have a problemSet page. If the set
is homework or jistr it will list the problems. If it is a gateway
it will provide a button to start a new version (if allowed) and
list all the current versions (visually displaed as attempts).

This simplifies the problemSets page since it no longer has to list
versions or link to different locations, since all sets now have
a set details page in problemSet.

Allow set headers to be edited for gateway assignments since the
set header is now visible in problemSet for gateway assignments.