Gateway quizzes should not always create a new version when attempts_per_version is 0.
Currently when attempts_per_version (Number of Graded Submissions per
Test) is 0 and you click on the gateway template in the problem sets
page, it always creates a new version even if there is a version
in progress with no submissions.  In this case, the in progress version
should be opened instead, the same as it does for any value greater than
zero.

The current logic checks if the current number of submissions is greater
than or equal to the max number of submissions or if the time has
expired.  If that happens then a new version is created.  So if the max
number of submissions is greater than zero then it does not create a new
version until all submissions are used up, which is how it presumably
was intended.

However, if the number of submissions is meant to be infinite, and so
the setting is 0, then the current number of submissions is always
greater than or equal to the setting of 0, and so it always creates a
new version even if no submissions have been made to an in progress
version.

This changes the logic so that if attempts_per_version is 0 (so an
infinite number of submissions is allowed), then at least one submission
is required before a new version is created.