Show rendering warnings in the library browser and homework set editor.
To make this work those warnings needed to be added to the json data
that is sent by formatRendereredProblem.

I also improve the appearance of these warnings with a bootstrap alert.

I removed the hardcoded warning header in the debug_format.pl.  The
warnings that are output already have headers, so that isn't really
needed.