Fix and improve some things on the new Gateway quiz ProblemSet.pm page.
The current implemenation assumes that when an un-versions quiz_mode
link is used, a new version of the Gateway quiz will be created, but
that is not always the case.

If a gateway quiz is set to allow more than one submission per version
(but not infinite), then the page shows "Start New Test" and presents
the un-versioned quiz_mode link after the first submission of the
current version.  The result is that when this button is clicked, the
current quiz with that one submission is opened instead.  Not a new
test.  This changes it so that the "Continue Open Test" button is shown
until all submissions are used up.  The only way to obtain the behavior
that was intended would be to modify the GatewayQuiz.pm module.

Also, the current implementation does not take the grace period into
account.  If the test is in the grace period, the GatewayQuiz.pm module
will also re-enter the currently open quiz if the un-versioned link is
used.  The page currently shows the "Start New Test" button in this case
as well, and using the button again results in that current version
being opened, and not a new test.  In this case this is changed so that
if the quiz is in the grace period, a message is displaying telling the
student this and what needs to be done to submit the test for a grade.
Of course the "Continue Open Test" button is also displayed in this
case.

Another issue that is fixed is that for accessibility purposes a
checkbox should always have a visible label.  The ProblemSets.pm page
uses the set name for this, and when the quiz versions were listed there
the quiz version was used for the label for those.  So that is done here
as well.

Another issue is that if the quiz is in the last minute, the message is
displayed as "You have remaining to complete the current test."  So this
changes it so that seconds are shown during the last minute.  Should
this be taken further and show minutes and seconds near the end of the
test?  Currently when there is 1 minute and 59 seconds left, the display
shows "1 minute".

The time remaining is removed from the "Test Versions" table.  That
information is already displayed above for the currently open version,
and so it isn't needed in the table.  Furthermore that display was also
plagued with more issues than the time remaining display above.  In the
last minute it shows "remain", and in the minute before that it shows "1
minute remain".

This also makes the messages at the top of the page stand out by using a
bootstrap alert.

The "#" is removed from the numbers in the test version list.  That
wasn't there when these were on the ProblemSets.pm page, and isn't
needed here.  Also, the `text-nowrap` class is added to prevent that
version number from going to the next line.

There is also a minor margin tweak for the message that shows "the
selected problem set is not valid" in GatewayQuiz.pm, and some other
minor clean up in ProblemSets.pm.